refactor(extension): extract shared executeBatch from duplicate batch runners
