+++ vscode-gotest/src/coverage.ts
@@ -7,7 +7,6 @@ import { tmpdir } from "node:os";
 import type { GoTestController } from "./testController.js";
 import type { DiscoveryCache } from "./discovery.js";
 import type { CoverageStore } from "./coverageStore.js";
-import { parseTestEvents, groupEventsByPackage } from "./outputParser.js";
 import {
   buildCliCommand,
   formatCliCommand,
@@ -17,11 +16,10 @@ import {
 import {
   collectItems,
   groupByPackage,
-  applyResults,
   spawnTestProcess,
   buildRunFilter,
-  computeWildcard,
 } from "./runnerUtils.js";
+import { executeBatch } from "./batchRunner.js";
 
 const execFileAsync = promisify(execFile);
 
@@ -526,146 +524,24 @@ export class CoverageRunner implements vscode.Disposable {
     token: vscode.CancellationToken,
     testOnly?: boolean,
   ): Promise<string> {
-    const importPaths = pkgInfos.map((p) => p.importPath);
-    const wildcard = filter ? undefined : computeWildcard(importPaths);
-    const cliPkgArgs = wildcard ? [wildcard] : importPaths;
-    let coverFile: string | undefined;
-
-    try {
-      const tmpDir = await mkdtemp(path.join(tmpdir(), "gotest-cov-"));
-      coverFile = path.join(tmpDir, "cover.out");
-
-      const cliArgs: string[] = [
-        "-json",
-        "-count=1",
-        "-covermode=atomic",
-        `-coverprofile=${coverFile}`,
-        ...cliPkgArgs,
-      ];
-      if (testOnly) {
-        cliArgs.push("-coverpkg=./...");
-      }
-      if (filter) {
-        cliArgs.push("-run", filter);
-      }
-      cliArgs.push(...testFlags);
-
-      const cmd = await buildCliCommand(
-        cliArgs,
-        workspaceDir,
-        this.outputChannel,
-      );
-      this.outputChannel.appendLine(`[coverage] ${formatCliCommand(cmd)}`);
-
-      const result = await spawnTestProcess(
-        cmd.bin,
-        cmd.args,
-        workspaceDir,
-        token,
-        this.outputChannel,
-        "coverage",
-      );
-
-      if (token.isCancellationRequested) {
-        for (const info of pkgInfos) {
-          for (const item of info.items) {
-            run.skipped(item);
-          }
+    const result = await executeBatch({
+      pkgInfos,
+      filter,
+      workspaceDir,
+      testFlags,
+      run,
+      token,
+      controller: this.controller,
+      outputChannel: this.outputChannel,
+      label: "coverage",
+      coverage: { store: this.store, testOnly },
+      onResults: (applied) => {
+        for (const r of applied) {
+          this.controller.recordResult(r.itemId, r.status, r.duration);
         }
-        return result.stdout;
-      }
-
-      const events = result.stdout ? parseTestEvents(result.stdout) : [];
-      const eventsByPkg = groupEventsByPackage(events);
-
-      for (const info of pkgInfos) {
-        const pkgEvents = eventsByPkg.get(info.importPath) ?? [];
-        if (pkgEvents.length > 0) {
-          const applied = applyResults(
-            this.controller,
-            run,
-            pkgEvents,
-            info.importPath,
-            info.dir,
-          );
-          for (const r of applied) {
-            this.controller.recordResult(r.itemId, r.status, r.duration);
-          }
-        }
-      }
-
-      if (result.exitCode !== 0) {
-        const diagnostic = [
-          result.stderr.trim(),
-          result.stdout.trim(),
-          `exit code ${result.exitCode}`,
-        ].filter(Boolean).join("\n\n");
-
-        for (const info of pkgInfos) {
-          if ((eventsByPkg.get(info.importPath) ?? []).length === 0) {
-            for (const item of info.items) {
-              run.errored(item, new vscode.TestMessage(diagnostic));
-            }
-          }
-        }
-      }
-      try {
-        const coverContent = await readFile(coverFile, "utf-8");
-        let funcOutput: string | undefined;
-        try {
-          funcOutput = await runGoToolCoverFunc(coverFile, workspaceDir);
-        } catch {
-          this.outputChannel.appendLine(
-            "[coverage] go tool cover -func failed, skipping declaration coverage",
-          );
-        }
-
-        if (testOnly) {
-          for (const info of pkgInfos) {
-            this.store.update(
-              info.importPath,
-              coverContent,
-              funcOutput,
-              true,
-            );
-          }
-        } else {
-          const coverByPkg = splitCoverByPackage(coverContent, importPaths);
-          const funcByPkg = funcOutput
-            ? splitFuncCoverageByPackage(funcOutput, importPaths)
-            : undefined;
-
-          for (const info of pkgInfos) {
-            const pkgCover = coverByPkg.get(info.importPath);
-            if (pkgCover) {
-              this.store.update(
-                info.importPath,
-                pkgCover,
-                funcByPkg?.get(info.importPath),
-              );
-            }
-          }
-        }
-      } catch {
-        this.outputChannel.appendLine("[coverage] no coverprofile generated");
-      }
-      return result.stdout;
-    } catch (err: unknown) {
-      const message = err instanceof Error ? err.message : String(err);
-      this.outputChannel.appendLine(`[coverage] error: ${message}`);
-      for (const info of pkgInfos) {
-        for (const item of info.items) {
-          run.errored(item, new vscode.TestMessage(message));
-        }
-      }
-      return "";
-    } finally {
-      if (coverFile) {
-        rm(path.dirname(coverFile), { recursive: true, force: true }).catch(
-          () => {},
-        );
-      }
-    }
+      },
+    });
+    return result.stdout;
   }
 
   async copyCoverageSummary(): Promise<void> {
